refactor: split AdminPanel into modular components
