+++ src/components/AdminPanel.jsx
@@ -8,9 +8,6 @@ import {
     updateEvent,
     deleteEvent,
     setAuthToken,
-    verifyBooking,
-    checkInBooking,
-    undoCheckIn,
     updateBooking,
     deleteBooking,
     deleteAllBookings,
@@ -29,7 +26,6 @@ import {
     ListItemIcon,
     ListItemText,
     TextField,
-    Autocomplete as MUIAutocomplete,
     Button,
     CssBaseline,
     Table,
@@ -87,25 +83,12 @@ import LockClockIcon from "@mui/icons-material/LockClock";
 import CheckCircleIcon from "@mui/icons-material/CheckCircle";
 import EditIcon from "@mui/icons-material/Edit";
 import ImageIcon from "@mui/icons-material/Image";
-import PlaceIcon from "@mui/icons-material/Place";
 import EuroIcon from "@mui/icons-material/Euro";
 import AccessTimeIcon from "@mui/icons-material/AccessTime";
 import MapIcon from "@mui/icons-material/Map";
-import CloseIcon from "@mui/icons-material/Close";
 import QrCodeScannerIcon from "@mui/icons-material/QrCodeScanner";
 import IosShareIcon from "@mui/icons-material/IosShare";
 import CalendarMonthIcon from "@mui/icons-material/CalendarMonth";
-import ErrorOutlineIcon from "@mui/icons-material/ErrorOutline";
-import WarningAmberIcon from "@mui/icons-material/WarningAmber";
-import LinkIcon from "@mui/icons-material/Link";
-import CircularProgress from "@mui/material/CircularProgress";
-import FullscreenIcon from "@mui/icons-material/Fullscreen";
-import FullscreenExitIcon from "@mui/icons-material/FullscreenExit";
-import VolumeUpIcon from "@mui/icons-material/VolumeUp";
-import VolumeOffIcon from "@mui/icons-material/VolumeOff";
-import StopCircleIcon from "@mui/icons-material/StopCircle";
-import UndoIcon from "@mui/icons-material/Undo";
-import CheckCircleOutlineIcon from "@mui/icons-material/CheckCircleOutline";
 
 
 
@@ -114,11 +97,14 @@ import ConfirmDialog from "./ConfirmDialog";
 import { useToast } from "./ToastContext";
 import { signOut } from "firebase/auth";
 import { auth } from "../firebase/config";
-import jsQR from "jsqr";
+import PlaceAutocomplete from "./admin/PlaceAutocomplete";
+import DriveImagePickerDialog from "./admin/DriveImagePickerDialog";
+import MobileEventCard from "./admin/MobileEventCard";
+import MobileBookingCard from "./admin/MobileBookingCard";
+import CheckInBox from "./admin/CheckInBox";
+import { isPast, eventDateTime } from "./admin/eventUtils";
 
 
-// Drive utils
-import { listImagesInFolder, driveCdnSrc, driveApiSrc } from "../lib/driveGallery";
 
 /* ---------------- THEME ---------------- */
 const drawerWidth = 256;
@@ -169,519 +155,16 @@ const toISO = (v) => {
 
 
 const todayISO = () => new Date().toISOString().slice(0, 10);
-const eventDateTime = (ev) => {
-    const d = ev?.date || "";
-    const t = ev?.time || "00:00";
-    const dt = new Date(`${d}T${t}:00`);
-    return isNaN(dt.getTime()) ? null : dt;
-};
-const isPast = (ev) => {
-    const dt = eventDateTime(ev);
-    return dt ? dt.getTime() < Date.now() : false;
-};
 
 // ======= GEOAPIFY Autocomplete (OSM) =======
 
-function PlaceAutocomplete({
-                               value,
-                               onChange,
-                               inputValue,
-                               onInputChange,
-                               error,
-                               helperText,
-                               onCoords,
-                           }) {
-    const [preds, setPreds] = useState([]);
-    const [loading, setLoading] = useState(false);
-    const [initError, setInitError] = useState("");
-    const abortRef = useRef(null);
-    const debounceRef = useRef(null);
-
-    const apiKey =
-        (import.meta?.env && import.meta.env.VITE_GEOAPIFY_KEY) ||
-        (window.APP_CONFIG && window.APP_CONFIG.GEOAPIFY_KEY);
-
-    useEffect(() => {
-        if (!apiKey) setInitError("Geoapify API key mancante");
-    }, [apiKey]);
-
-    useEffect(() => {
-        const q = (inputValue || "").trim();
-        if (!apiKey) return;
-        if (!q) {
-            setPreds([]);
-            return;
-        }
-        setLoading(true);
-        if (debounceRef.current) clearTimeout(debounceRef.current);
-        debounceRef.current = setTimeout(async () => {
-            try {
-                if (abortRef.current) abortRef.current.abort();
-                abortRef.current = new AbortController();
-
-                const url = new URL("https://api.geoapify.com/v1/geocode/autocomplete");
-                url.searchParams.set("text", q);
-                url.searchParams.set("limit", "7");
-                url.searchParams.set("lang", "it");
-                // opzionale: restrizione paese → Italia
-                // url.searchParams.set("filter", "countrycode:it");
-                url.searchParams.set("apiKey", apiKey);
-
-                const r = await fetch(url.toString(), { signal: abortRef.current.signal });
-                const data = await r.json();
-                const rows = (data?.features || []).map((f) => {
-                    const p = f.properties || {};
-                    return {
-                        label: p.formatted || p.address_line1 || p.name || "",
-                        place_id: p.place_id || p.datasource?.raw?.osm_id || p.osm_id || p.datasource?.feature_id || Math.random().toString(36).slice(2),
-                        main_text: p.address_line1 || p.name || p.street || "",
-                        secondary_text: p.address_line2 || [p.postcode, p.city, p.country].filter(Boolean).join(" "),
-                        lat: p.lat,
-                        lon: p.lon,
-                    };
-                });
-                setPreds(rows);
-            } catch (e) {
-                if (e?.name !== "AbortError") {
-                    setPreds([]);
-                }
-            } finally {
-                setLoading(false);
-            }
-        }, 220);
-
-        return () => {
-            if (debounceRef.current) clearTimeout(debounceRef.current);
-            if (abortRef.current) abortRef.current.abort();
-        };
-    }, [inputValue, apiKey]);
-
-    const selectOption = (_evt, opt) => {
-        if (!opt) {
-            onChange(null);
-            onCoords?.(null);
-            return;
-        }
-        onChange({
-            label: opt.label,
-            place_id: opt.place_id,
-            verified: true,
-            lat: opt.lat,
-            lon: opt.lon,
-        });
-        if (opt.lat && opt.lon) onCoords?.({ lat: opt.lat, lon: opt.lon });
-    };
-
-    return (
-        <MUIAutocomplete
-            options={preds}
-            value={value}
-            onChange={selectOption}
-            inputValue={inputValue}
-            onInputChange={(e, v) => onInputChange(v)}
-            getOptionLabel={(o) => o?.label || ""}
-            noOptionsText={initError ? "Geoapify non configurato" : "Nessun risultato"}
-            loading={loading}
-            clearOnBlur={false}
-            blurOnSelect
-            isOptionEqualToValue={(o, v) => o.place_id === v.place_id}
-            renderOption={(props, option) => (
-                <li {...props} key={option.place_id}>
-                    <Stack>
-                        <Typography variant="body2" fontWeight={600}>
-                            {option.main_text || option.label}
-                        </Typography>
-                        {option.secondary_text && (
-                            <Typography variant="caption" sx={{ opacity: 0.8 }}>
-                                {option.secondary_text}
-                            </Typography>
-                        )}
-                    </Stack>
-                </li>
-            )}
-            renderInput={(params) => {
-                const { InputProps, ...rest } = params;
-                return (
-                    <TextField
-                        {...rest}
-                        label="Cerca e seleziona un luogo"
-                        required
-                        error={!!error || !!initError}
-                        helperText={initError || helperText || "Suggerimenti da OSM/Geoapify"}
-                        InputProps={{
-                            ...InputProps,
-                            startAdornment: (
-                                <InputAdornment position="start">
-                                    <PlaceIcon fontSize="small" />
-                                </InputAdornment>
-                            ),
-                        }}
-                        sx={{
-                            "& .MuiOutlinedInput-root": {
-                                "& fieldset": { borderColor: "rgba(255,255,255,0.18)" },
-                                "&:hover fieldset": { borderColor: "primary.main" },
-                                "&.Mui-focused fieldset": { borderColor: "primary.main" },
-                            },
-                        }}
-                    />
-                );
-            }}
-        />
-    );
-}
 
 /* =========================================
    Dialog selezione copertina da Drive — griglia responsive (ottimizzata + HEIC)
    ========================================= */
-function DriveImagePickerDialog({ open, onClose, onPick }) {
-    const apiKey =
-        (import.meta?.env && import.meta.env.VITE_GOOGLE_API_KEY) ||
-        (window.APP_CONFIG && window.APP_CONFIG.GOOGLE_API_KEY);
-    const folderId =
-        (import.meta?.env && import.meta.env.VITE_GOOGLE_DRIVE_FOLDER_ID) ||
-        (window.APP_CONFIG && window.APP_CONFIG.GOOGLE_DRIVE_FOLDER_ID);
-
-    const [items, setItems] = useState([]);
-    const [loading, setLoading] = useState(false);
-    const [filter, setFilter] = useState("");
-    const [error, setError] = useState("");
-
-    // Preconnect al CDN immagini di Google per ridurre il TTFB
-    useEffect(() => {
-        const link = document.createElement("link");
-        link.rel = "preconnect";
-        link.href = "https://lh3.googleusercontent.com";
-        document.head.appendChild(link);
-        return () => { try { document.head.removeChild(link); } catch {} };
-    }, []);
-
-    const load = useCallback(async () => {
-        if (!apiKey || !folderId) {
-            setError("Config mancante: API_KEY o FOLDER_ID");
-            return;
-        }
-        setError("");
-        setLoading(true);
-        try {
-            const list = await listImagesInFolder(folderId, {
-                apiKey,
-                includeSharedDrives: true,
-                pageSize: 200,
-            });
-            setItems(Array.isArray(list) ? list : []);
-        } catch (e) {
-            console.error(e);
-            setError("Impossibile caricare le immagini da Drive.");
-        } finally {
-            setLoading(false);
-        }
-    }, [apiKey, folderId]);
-
-    useEffect(() => { if (open) load(); }, [open, load]);
-
-    const filtered = useMemo(() => {
-        const t = filter.trim().toLowerCase();
-        return t ? items.filter((i) => (i.name || "").toLowerCase().includes(t)) : items;
-    }, [filter, items]);
-
-    const [visible, setVisible] = useState(60);
-    const sentinelRef = useRef(null);
-
-    useEffect(() => { setVisible(60); }, [open, filter, items.length]);
-
-    useEffect(() => {
-        if (!open) return;
-        const el = sentinelRef.current;
-        if (!el) return;
-        const io = new IntersectionObserver((entries) => {
-            if (entries[0].isIntersecting) {
-                setVisible((v) => Math.min(v + 60, filtered.length));
-            }
-        }, { root: null, rootMargin: "1200px 0px", threshold: 0 });
-        io.observe(el);
-        return () => io.disconnect();
-    }, [open, filtered.length]);
-
-    /* ====== Thumb ottimizzato con lazy + srcset + HEIC fallback ====== */
-
-    // Cache per gli oggetti URL (HEIC → WEBP) per non riconvertire
-    const heicUrlCacheRef = useRef(new Map());
-
-    function DriveThumb({ id, name, mimeType, index }) {
-        const boxRef = useRef(null);
-        const imgRef = useRef(null);
-        const objectUrlRef = useRef(null);
-        const isHeic = useMemo(
-            () =>
-                /heic|heif/i.test(mimeType || "") ||
-                /\.heic$|\.heif$/i.test(name || ""),
-            [mimeType, name]
-        );
-
-        // Lazy load on-intersect
-        useEffect(() => {
-            const el = boxRef.current;
-            if (!el) return;
-            const io = new IntersectionObserver(
-                (entries) => {
-                    if (!entries[0].isIntersecting) return;
-                    const img = imgRef.current;
-                    if (!img || img.getAttribute("src")) return;
-
-                    // srcset: lascia decidere al browser la risoluzione
-                    const srcset = [200, 320, 480, 640, 800]
-                        .map((w) => `${driveCdnSrc(id, w)} ${w}w`)
-                        .join(", ");
-
-                    img.src = driveCdnSrc(id, 320);
-                    img.srcset = srcset;
-                    img.sizes =
-                        "(max-width:600px) 50vw, (max-width:900px) 33vw, (max-width:1200px) 25vw, 16.66vw";
-                    img.setAttribute("fetchpriority", index < 8 ? "high" : "low");
-
-                    io.disconnect();
-                },
-                { rootMargin: "300px 0px" }
-            );
-            io.observe(el);
-            return () => io.disconnect();
-        }, [id, index]);
-
-        // Prefetch versione grande per click istantaneo
-        const prefetchLarge = useCallback(() => {
-            const i = new Image();
-            i.decoding = "async";
-            i.src = driveCdnSrc(id, 1600);
-        }, [id]);
-
-        // Cleanup di eventuali objectURL creati per HEIC
-        const revokeObjUrl = useCallback(() => {
-            if (objectUrlRef.current) {
-                URL.revokeObjectURL(objectUrlRef.current);
-                objectUrlRef.current = null;
-            }
-        }, []);
-
-        useEffect(() => revokeObjUrl, [revokeObjUrl]);
-
-        const handleError = useCallback(async (e) => {
-            const img = e.currentTarget;
-
-            // Se già abbiamo una conversione in cache → usa quella
-            const cached = heicUrlCacheRef.current.get(id);
-            if (cached) {
-                img.removeAttribute("srcset");
-                img.removeAttribute("sizes");
-                img.src = cached;
-                return;
-            }
-
-            // Se HEIC/HEIF: prova transcodifica client-side → WEBP
-            if (isHeic) {
-                try {
-                    // prendi l'originale (potrebbe essere HEIC)
-                    const r = await fetch(driveApiSrc(id, apiKey));
-                    const blob = await r.blob();
-
-                    // import dinamico: caricato SOLO quando serve
-                    const heic2any = (await import("heic2any")).default;
-                    const out = await heic2any({ blob, toType: "image/webp", quality: 0.8 });
-
-                    const webpBlob = Array.isArray(out) ? out[0] : out;
-                    const url = URL.createObjectURL(webpBlob);
-                    objectUrlRef.current = url;
-                    heicUrlCacheRef.current.set(id, url);
-
-                    img.removeAttribute("srcset");
-                    img.removeAttribute("sizes");
-                    img.src = url;
-                    return;
-                } catch (err) {
-                    // se conversione fallisce, tenta comunque l'originale (se il browser lo supporta)
-                    console.warn("HEIC conversion failed, fallback to original", err);
-                }
-            }
-
-            // Fallback generico: original via API (PNG/JPEG ok, HEIC no se browser non supporta)
-            img.removeAttribute("srcset");
-            img.removeAttribute("sizes");
-            img.src = driveApiSrc(id, apiKey);
-        }, [apiKey, id, isHeic]);
-
-        return (
-            <Box
-                ref={boxRef}
-                role="button"
-                onClick={() => { onPick(driveCdnSrc(id, 1600)); onClose(); }}
-                onMouseEnter={prefetchLarge}
-                onTouchStart={prefetchLarge}
-                sx={{
-                    position: "relative",
-                    width: "100%",
-                    pt: "66.666%",
-                    overflow: "hidden",
-                    borderRadius: 1,
-                    border: "1px solid rgba(255,255,255,0.08)",
-                    cursor: "pointer",
-                    "&:hover": { outline: "2px solid", outlineColor: "primary.main" },
-                    contentVisibility: "auto",
-                    containIntrinsicSize: "120px 80px",
-                    background: "#111",
-                }}
-            >
-                <img
-                    ref={imgRef}
-                    alt={name}
-                    decoding="async"
-                    loading="lazy"
-                    onError={handleError}
-                    onLoad={() => {
-                        // se abbiamo usato un objectURL per HEIC, nulla da fare qui
-                    }}
-                    style={{
-                        position: "absolute",
-                        inset: 0,
-                        width: "100%",
-                        height: "100%",
-                        objectFit: "cover",
-                        display: "block",
-                    }}
-                />
-            </Box>
-        );
-    }
-
-    return (
-        <Dialog open={open} onClose={onClose} fullWidth maxWidth="lg" PaperProps={{ sx: { borderRadius: 2 } }}>
-            <DialogTitle sx={{ display: "flex", alignItems: "center", gap: 1 }}>
-                <ImageIcon color="primary" /> Seleziona copertina da Drive
-                <IconButton onClick={onClose} sx={{ ml: "auto" }}>
-                    <CloseIcon />
-                </IconButton>
-            </DialogTitle>
-
-            <DialogContent dividers sx={{ background: "#101014" }}>
-                <Stack direction={{ xs: "column", sm: "row" }} spacing={1} sx={{ mb: 2 }}>
-                    <TextField
-                        size="small"
-                        placeholder="Filtra per nome…"
-                        value={filter}
-                        onChange={(e) => setFilter(e.target.value)}
-                        InputProps={{
-                            startAdornment: (
-                                <InputAdornment position="start">
-                                    <SearchIcon fontSize="small" />
-                                </InputAdornment>
-                            ),
-                        }}
-                    />
-                    <Button onClick={load} variant="outlined" sx={{ color: "primary.main", borderColor: "primary.main" }}>
-                        Ricarica
-                    </Button>
-                </Stack>
-
-                {loading && (
-                    <Grid container spacing={1}>
-                        {Array.from({ length: 12 }).map((_, i) => (
-                            <Grid item xs={6} sm={4} md={3} key={i}>
-                                <Skeleton variant="rectangular" height={120} />
-                            </Grid>
-                        ))}
-                    </Grid>
-                )}
-
-                {!loading && error && (
-                    <Paper sx={{ p: 2, textAlign: "center", border: "1px dashed rgba(255,255,255,.15)" }}>{error}</Paper>
-                )}
-
-                {!loading && !error && (
-                    <Box
-                        sx={{
-                            display: "grid",
-                            gridTemplateColumns: {
-                                xs: "repeat(2, 1fr)",
-                                sm: "repeat(3, 1fr)",
-                                md: "repeat(4, 1fr)",
-                                lg: "repeat(6, 1fr)",
-                            },
-                            gap: 1,
-                        }}
-                    >
-                        {filtered.slice(0, visible).map((img, i) => (
-                            <DriveThumb
-                                key={img.id}
-                                id={img.id}
-                                name={img.name}
-                                mimeType={img.mimeType}
-                                index={i}
-                            />
-                        ))}
-                        <Box ref={sentinelRef} sx={{ height: 1, gridColumn: "1 / -1" }} />
-                        {!filtered.length && (
-                            <Box sx={{ gridColumn: "1/-1", p: 3, textAlign: "center", opacity: 0.7 }}>
-                                Nessuna immagine trovata.
-                            </Box>
-                        )}
-                    </Box>
-                )}
-            </DialogContent>
-        </Dialog>
-    );
-}
 
 
 /* ------------ MOBILE CARDS ------------ */
-function MobileEventCard({ ev, onEdit, onDelete, onToggleSoldOut, onDuplicate, onExportICS, canDelete }) {
-    const past = isPast(ev);
-    const statusColor =
-        ev.status === "draft" ? "default" : ev.status === "archived" ? "warning" : "success";
-    return (
-        <Card sx={{ mb: 1.5, background: "#18181f", border: "1px solid rgba(255,255,255,0.06)" }}>
-            <CardContent sx={{ pb: 1.5 }}>
-                <Stack direction="row" spacing={1} alignItems="center" sx={{ mb: 0.5, flexWrap: "wrap" }}>
-                    {!past ? <CheckCircleIcon color="success" fontSize="small" /> : <LockClockIcon color="disabled" fontSize="small" />}
-                    <Typography variant="subtitle1" fontWeight={700}>{ev.name}</Typography>
-                    <Chip size="small" label={ev.status || "published"} sx={{ ml: "auto" }} color={statusColor} />
-                </Stack>
-                <Typography variant="body2" sx={{ opacity: 0.85 }}>{ev.dj || "—"}</Typography>
-                <Typography variant="body2" sx={{ opacity: 0.85 }}>{ev.place || "—"}</Typography>
-                <Stack direction="row" spacing={2} sx={{ mt: 1, opacity: 0.8 }}>
-                    <Typography variant="caption">📅 {ev.date}</Typography>
-                    <Typography variant="caption">⏰ {ev.time}</Typography>
-                    <Typography variant="caption">👥 {ev.capacity || "-"}</Typography>
-                </Stack>
-            </CardContent>
-            <CardActions sx={{ pt: 0, pb: 1.5, px: 2, justifyContent: "space-between" }}>
-                <Stack direction="row" spacing={1} alignItems="center">
-                    <Typography variant="caption">Sold Out</Typography>
-                    <Switch size="small" color="warning" checked={!!ev.soldOut} onChange={(e) => onToggleSoldOut(e.target.checked)} disabled={past} />
-                </Stack>
-                <Stack direction="row" spacing={1}>
-                    <Tooltip title="Duplica"><IconButton size="small" onClick={onDuplicate}><ContentCopyIcon fontSize="small" /></IconButton></Tooltip>
-                    <Tooltip title=".ics"><IconButton size="small" onClick={onExportICS}><CalendarMonthIcon fontSize="small" /></IconButton></Tooltip>
-                    <Button size="small" startIcon={<EditIcon />} onClick={onEdit} disabled={past}>Modifica</Button>
-                    <span>
-            <IconButton size="small" color="error" onClick={onDelete} disabled={past || !canDelete}><DeleteIcon fontSize="small" /></IconButton>
-          </span>
-                </Stack>
-            </CardActions>
-        </Card>
-    );
-}
-function MobileBookingCard({ b }) {
-    return (
-        <Card sx={{ mb: 1.2, background: "#18181f", border: "1px solid rgba(255,255,255,0.06)" }}>
-            <CardContent>
-                <Typography variant="subtitle1" fontWeight={700}>{b.nome} {b.cognome}</Typography>
-                <Typography variant="body2" sx={{ opacity: 0.85 }}>{b.email}</Typography>
-                <Stack direction="row" spacing={2} sx={{ mt: 0.5, opacity: 0.85 }}>
-                    <Typography variant="caption">📞 {b.telefono}</Typography>
-                    <Typography variant="caption">🎟️ {b.quantity || 1}</Typography>
-                </Stack>
-            </CardContent>
-        </Card>
-    );
-}
 
 /* ---------- Helpers ---------- */
 const roles = { admin: "admin", editor: "editor", staff: "staff" };
@@ -2566,686 +2049,5 @@ const AdminPanel = () => {
         </MuiThemeProvider>
     );
 };
-
-
-/* =======================
-   Check-in component
-   ======================= */
-function CheckInBox({ events = [] }) {
-    const videoRef = useRef(null);
-    const canvasRef = useRef(null);
-    const trackRef = useRef(null);
-    const wakeLockRef = useRef(null);
-
-    const [support, setSupport] = useState({ camera: false, detector: false });
-    const [scanning, setScanning] = useState(false);
-    const [verifying, setVerifying] = useState(false);
-    const [lastScan, setLastScan] = useState("");
-    const [result, setResult] = useState(null); // { valid, reason, token, ... }
-    const [onlyToday, setOnlyToday] = useState(true);
-    const [opOpen, setOpOpen] = useState(false); // modal operatore
-    const [manual, setManual] = useState("");
-
-    const { showToast } = useToast();
-
-    const todayISO = new Date().toISOString().slice(0,10);
-    const eventById = useCallback((id) => events.find((e) => e.id === id), [events]);
-
-    // Audio
-    const audioCtxRef = useRef(null);
-    const [soundOn, setSoundOn] = useState(() => localStorage.getItem("checkinSound") !== "off");
-
-    // dentro CheckInBox, in alto tra gli hook:
-    const theme = useTheme();
-    const isXs = useMediaQuery(theme.breakpoints.down("sm"));
-
-
-// sblocca l’audio su gesto utente (iOS richiede interazione)
-    async function ensureAudio() {
-        try {
-            if (!audioCtxRef.current) {
-                const ACtx = window.AudioContext || window.webkitAudioContext;
-                audioCtxRef.current = new ACtx();
-            }
-            if (audioCtxRef.current?.state === "suspended") {
-                await audioCtxRef.current.resume();
-            }
-        } catch {}
-    }
-
-    function beep({ freq = 880, when = 0, duration = 0.12, type = "sine", gain = 0.03 }) {
-        const ctx = audioCtxRef.current;
-        if (!ctx) return;
-        const t0 = ctx.currentTime + when;
-        const osc = ctx.createOscillator();
-        const g = ctx.createGain();
-        osc.type = type;
-        osc.frequency.setValueAtTime(freq, t0);
-        g.gain.setValueAtTime(gain, t0);
-        g.gain.exponentialRampToValueAtTime(0.0001, t0 + duration);
-        osc.connect(g).connect(ctx.destination);
-        osc.start(t0);
-        osc.stop(t0 + duration + 0.02);
-    }
-
-// “ding” (successo)
-    function playDing() {
-        if (!soundOn) return;
-        ensureAudio();
-        beep({ freq: 880, duration: 0.11, gain: 0.04 });
-        beep({ freq: 1320, when: 0.12, duration: 0.10, gain: 0.04 });
-    }
-
-// “beep-beep” (warning)
-    function playWarn() {
-        if (!soundOn) return;
-        ensureAudio();
-        beep({ freq: 650, duration: 0.08, gain: 0.035 });
-        beep({ freq: 650, when: 0.18, duration: 0.08, gain: 0.035 });
-    }
-
-// “buzzer” (errore)
-    function playError() {
-        if (!soundOn) return;
-        ensureAudio();
-        const ctx = audioCtxRef.current;
-        if (!ctx) return;
-        const t0 = ctx.currentTime;
-        const osc = ctx.createOscillator();
-        const g = ctx.createGain();
-        osc.type = "square";
-        osc.frequency.setValueAtTime(360, t0);
-        osc.frequency.exponentialRampToValueAtTime(180, t0 + 0.25);
-        g.gain.setValueAtTime(0.035, t0);
-        g.gain.exponentialRampToValueAtTime(0.0001, t0 + 0.25);
-        osc.connect(g).connect(ctx.destination);
-        osc.start(t0);
-        osc.stop(t0 + 0.27);
-    }
-
-
-
-    useEffect(() => {
-        setSupport({
-            camera: !!navigator.mediaDevices?.getUserMedia,
-            detector: "BarcodeDetector" in window,
-        });
-    }, []);
-
-    async function requestWakeLock() { try { wakeLockRef.current = await navigator.wakeLock?.request("screen"); } catch {} }
-    function releaseWakeLock() { try { wakeLockRef.current?.release(); wakeLockRef.current = null; } catch {} }
-
-    useEffect(() => {
-        const onVis = () => document.hidden && stopScan();
-        document.addEventListener("visibilitychange", onVis);
-        return () => document.removeEventListener("visibilitychange", onVis);
-    }, []);
-
-    const extractToken = (text) => {
-        try { const u = new URL(text); const tk = u.searchParams.get("token"); if (tk) return tk; } catch {}
-        const maybeJwt = (text || "").trim();
-        if (maybeJwt.split(".").length === 3) return maybeJwt;
-        const m = text.match(/[?&]token=([^&\s]+)/i);
-        if (m) return m[1];
-        return null;
-    };
-
-    const doVerify = async (text, opts = { autoCheckIn: true }) => {
-        const token = extractToken(text);
-        if (!token) { showToast("QR/Link non valido: nessun token trovato", "error"); return; }
-        setLastScan(text);
-        setVerifying(true);
-        try {
-            const data = await verifyBooking(token);
-            const ev = eventById(data.eventId);
-            const isToday = ev?.date === todayISO;
-            const payload = { ...data, token, isToday };
-            setResult(payload);
-
-            setOpOpen(true); // apri modal operatore
-
-            if (!data.valid) { showToast("Token non valido", "error"); return; }
-            if (onlyToday && !isToday) { showToast("Evento non è oggi — verifica manuale", "warning"); return; }
-
-            // ✅ AUTO CHECK-IN: +1 alla scansione
-            if (opts.autoCheckIn && data.remaining > 0) {
-                await doCheckIn(token, data.remaining, { silent: true });
-            } else if (data.remaining <= 0) {
-                showToast("Prenotazione già esaurita", "error");
-                playWarn();
-            }
-        } catch {
-            setResult({ valid: false, reason: "network_error" });
-            showToast("Errore di verifica", "error");
-        } finally {
-            setVerifying(false);
-        }
-    };
-
-    const stopScan = () => {
-        setScanning(false);
-        const stream = videoRef.current?.srcObject;
-        if (stream) { stream.getTracks().forEach((t) => t.stop()); videoRef.current.srcObject = null; }
-        trackRef.current = null;
-        releaseWakeLock();
-    };
-
-    const startScan = async () => {
-        if (!support.camera) return;
-        setResult(null);
-        setScanning(true);
-        await requestWakeLock();
-        await ensureAudio();
-        try {
-            const stream = await navigator.mediaDevices.getUserMedia({
-                video: { facingMode: { ideal: "environment" }, width: { ideal: 1280 }, height: { ideal: 720 } },
-                audio: false,
-            });
-            if (videoRef.current) { videoRef.current.srcObject = stream; await videoRef.current.play(); }
-            trackRef.current = stream.getVideoTracks()[0];
-
-            await new Promise((r) => {
-                const v = videoRef.current;
-                const check = () => (v && v.videoWidth > 0) ? r() : setTimeout(check, 60);
-                check();
-            });
-
-            const v = videoRef.current;
-            const c = canvasRef.current;
-            const scale = Math.min(640 / Math.max(v.videoWidth, v.videoHeight), 1);
-            c.width  = Math.round(v.videoWidth * scale);
-            c.height = Math.round(v.videoHeight * scale);
-            const ctx = c.getContext("2d", { willReadFrequently: true });
-
-            let detector = null;
-            if (support.detector) { try { detector = new window.BarcodeDetector({ formats: ["qr_code"] }); } catch { detector = null; } }
-
-            const loop = async () => {
-                if (!videoRef.current || !scanning) return;
-                try {
-                    if (detector) {
-                        const detections = await detector.detect(videoRef.current);
-                        if (detections?.length) { const text = detections[0].rawValue; stopScan(); await doVerify(text, { autoCheckIn: true }); return; }
-                    }
-                    ctx.drawImage(v, 0, 0, c.width, c.height);
-                    const imageData = ctx.getImageData(0, 0, c.width, c.height);
-                    const qr = jsQR(imageData.data, imageData.width, imageData.height, { inversionAttempts: "attemptBoth" });
-                    if (qr?.data) { stopScan(); await doVerify(qr.data, { autoCheckIn: true }); return; }
-                } catch {}
-                requestAnimationFrame(loop);
-            };
-            loop();
-        } catch {
-            setScanning(false);
-            showToast("Impossibile accedere alla fotocamera", "error");
-            releaseWakeLock();
-        }
-    };
-
-    const onPickImage = async (file) => {
-        if (!file) return;
-        const img = new Image();
-        img.onload = async () => {
-            const c = canvasRef.current;
-            const ctx = c.getContext("2d", { willReadFrequently: true });
-            c.width = img.width; c.height = img.height;
-            ctx.drawImage(img, 0, 0);
-            const imageData = ctx.getImageData(0, 0, c.width, c.height);
-            const qr = jsQR(imageData.data, imageData.width, imageData.height, { inversionAttempts: "attemptBoth" });
-            if (qr?.data) await doVerify(qr.data, { autoCheckIn: true });
-            else showToast("Nessun QR riconosciuto", "error");
-        };
-        img.onerror = () => showToast("Immagine non valida", "error");
-        img.src = URL.createObjectURL(file);
-    };
-
-    // AZIONI
-    const doCheckIn = async (token, count = 1, { silent = false } = {}) => {
-        try {
-            const r = await checkInBooking(token, count);
-            setResult(prev => prev ? { ...prev, checkedInCount: r.checkedInCount, remaining: r.remaining } : prev);
-            if (!silent) showToast("Check-in registrato ", "success");
-            if (!silent) playDing();
-            else playDing(); // anche per l’auto +1 va bene dare feedback
-
-            try { navigator.vibrate?.(30); } catch {}
-        } catch (e) {
-            const code = e?.response?.data?.error || e?.message;
-            if (code === "already_fully_checked_in") showToast("Prenotazione già esaurita", "error");
-            else if (code === "exceeds_quantity") showToast("Numero superiore ai biglietti disponibili", "error");
-            else if (code === "token_mismatch") showToast("Token non corrisponde alla prenotazione", "error");
-            else showToast("Errore nel check-in", "error");
-            playError();
-        }
-    };
-
-    const doUndo = async () => {
-        if (!result?.token) return;
-        try {
-            const r = await undoCheckIn(result.token, 1);
-            setResult(prev => prev ? { ...prev, checkedInCount: r.checkedInCount, remaining: r.remaining } : prev);
-            showToast("Ultimo check-in annullato", "success");
-            try { navigator.vibrate?.([20, 20]); } catch {}
-        } catch (e) {
-            const code = e?.response?.data?.error || e?.message;
-            if (code === "nothing_to_undo") showToast("Nulla da annullare", "info");
-            else showToast("Errore nell'annullamento", "error");
-        }
-    };
-
-    // VIEW helpers
-    const statusKind = (() => {
-        if (!result) return "idle";
-        if (!result.valid) return result.reason === "expired" ? "warn" : "error";
-        if (onlyToday && result.valid && !result.isToday) return "warn";
-        if (result.valid && result.remaining > 0) return "ok";
-        if (result.valid && result.remaining === 0) return "done";
-        return "error";
-    })();
-
-    const colorByKind = {
-        ok:   { bg: "#10391F", border: "#2ECC71" },
-        done: { bg: "#102537", border: "#5DADE2" },
-        warn: { bg: "#3A2F10", border: "#F1C40F" },
-        error:{ bg: "#3A1010", border: "#E74C3C" },
-        idle: { bg: "#1a1a22", border: "#444" },
-    };
-
-    const ev = result?.eventId ? eventById(result.eventId) : null;
-
-    return (
-        <Box>
-            {/* Controlli rapidi */}
-            <Stack direction="row" spacing={1} alignItems="center" sx={{ mb: 1.5 }} flexWrap="wrap">
-                <FormControl>
-                    <Stack direction="row" spacing={1} alignItems="center">
-                        <Switch checked={onlyToday} onChange={(_, v) => setOnlyToday(v)} />
-                        <FormHelperText sx={{ color: "inherit" }}>Solo eventi di oggi ({todayISO})</FormHelperText>
-                    </Stack>
-                </FormControl>
-                <Button variant="outlined" startIcon={<FullscreenIcon />} onClick={() => setOpOpen(true)}>
-                    Modal operatore
-                </Button>
-            </Stack>
-
-            <Grid container spacing={2}>
-                {/* Scanner */}
-                <Grid item xs={12} md={6}>
-                    <Paper
-                        sx={{
-                            p: { xs: 1.25, md: 1.5 },       // 10px mobile
-                            borderRadius: 2,
-                            border: "1px solid rgba(255,255,255,0.12)",
-                            background: "#0c0c12",
-                        }}
-                    >
-
-                    <Typography variant="subtitle1" sx={{ mb: 1, fontWeight: 700 }}>Scanner QR</Typography>
-                        <Box sx={{ position: "relative", width: "100%", borderRadius: 2, overflow: "hidden", border: "1px solid rgba(255,255,255,0.15)", background: "#000" }}>
-                            <video ref={videoRef} playsInline muted style={{ width: "100%", height: "auto", display: "block" }} />
-                            {!scanning && (
-                                <Box sx={{ position: "absolute", inset: 0, display: "grid", placeItems: "center", p: 2,
-                                    background: "repeating-linear-gradient(45deg, rgba(255,255,255,0.03) 0 10px, rgba(255,255,255,0.06) 10px 20px)" }}>
-                                    <Stack spacing={1} alignItems="center">
-                                        <QrCodeScannerIcon sx={{ fontSize: 42, opacity: 0.7 }} />
-                                        <Typography variant="body2" sx={{ opacity: 0.8, textAlign: "center" }}>
-                                            Premi “Avvia scansione” e inquadra il QR
-                                        </Typography>
-                                    </Stack>
-                                </Box>
-                            )}
-                        </Box>
-
-                        {/* --- PATCH mobile: pulsanti scanner --- */}
-                        <Stack
-                            direction={{ xs: "column", sm: "row" }}
-                            spacing={1}
-                            sx={{ mt: 1, "& .MuiButton-root": { minHeight: 44 } }}
-                        >
-                            {/* riga 1 su mobile: Avvia + Carica (50/50) */}
-                            <Stack direction={{ xs: "column", md: "row" }} spacing={1}>
-                                <Button
-                                    startIcon={<QrCodeScannerIcon />}
-                                    variant="contained"
-                                    onClick={startScan}
-                                    disabled={!support.camera || scanning}
-                                    sx={{ width: { xs: "100%", md: "auto" } }}
-                                >
-                                    Avvia scansione
-                                </Button>
-
-                                <Button
-                                    variant="outlined"
-                                    component="label"
-                                    sx={{ width: { xs: "100%", md: "auto" } }}
-                                >
-                                    Carica immagine
-                                    <input
-                                        hidden
-                                        accept="image/*"
-                                        type="file"
-                                        onChange={(e) => onPickImage(e.target.files?.[0])}
-                                    />
-                                </Button>
-                            </Stack>
-
-
-                            {/* riga 2 su mobile: Stop + volume */}
-                            <Stack direction="row" spacing={1} sx={{ width: "100%" }}>
-                                <Button
-                                    variant="text"
-                                    onClick={stopScan}
-                                    disabled={!scanning}
-                                    sx={{ flex: { xs: "1 1 100%", sm: "0 0 auto" } }}
-                                >
-                                    Stop
-                                </Button>
-
-                                <IconButton
-                                    onClick={() => {
-                                        const next = !soundOn;
-                                        setSoundOn(next);
-                                        try { localStorage.setItem("checkinSound", next ? "on" : "off"); } catch {}
-                                    }}
-                                    aria-label={soundOn ? "Disattiva suoni" : "Attiva suoni"}
-                                    sx={{ ml: { xs: 0, sm: "auto" } }}
-                                >
-                                    {soundOn ? <VolumeUpIcon /> : <VolumeOffIcon />}
-                                </IconButton>
-                            </Stack>
-                        </Stack>
-
-
-                        <FormHelperText sx={{ mt: 1 }}>
-                            In alternativa carica una foto del QR.
-                        </FormHelperText>
-
-                        {/* Verifica manuale (link o token) */}
-                        <Box sx={{ mt: 1.5 }}>
-                            <Typography variant="subtitle2" sx={{ mb: 0.5, opacity: 0.9 }}>
-                                Verifica manuale (link del QR o token JWT)
-                            </Typography>
-                            {/* --- PATCH mobile: verifica manuale --- */}
-                            <Stack direction={{ xs: "column", sm: "row" }} spacing={1}>
-                                <TextField
-                                    placeholder="Incolla link completo ricevuto via email, o il token…"
-                                    value={manual}
-                                    onChange={(e) => setManual(e.target.value)}
-                                    onKeyDown={(e) => {
-                                        if (e.key === "Enter" && manual && !verifying) {
-                                            e.preventDefault();
-                                            doVerify(manual, { autoCheckIn: true });
-                                        }
-                                    }}
-                                    size="small"
-                                    fullWidth
-                                />
-                                <Button
-                                    variant="contained"
-                                    onClick={() => doVerify(manual, { autoCheckIn: true })}
-                                    disabled={!manual || verifying}
-                                    sx={{ width: { xs: "100%", sm: "auto" } }}
-                                >
-                                    {verifying ? <CircularProgress size={18} /> : "Verifica"}
-                                </Button>
-                            </Stack>
-
-                            <FormHelperText sx={{ mt: 0.5 }}>
-                                Se valido e l’evento è di oggi (toggle attivo), viene registrato automaticamente un +1 check-in.
-                            </FormHelperText>
-                        </Box>
-
-
-                        <canvas ref={canvasRef} style={{ display: "none" }} />
-                    </Paper>
-                </Grid>
-
-                {/* Esito & azioni */}
-                <Grid item xs={12} md={6}>
-                    <Paper
-                        sx={{
-                            p: { xs: 1.25, md: 1.5 },
-                            borderRadius: 2,
-                            border: {
-                                xs: `1px solid ${colorByKind[statusKind].border}`,
-                                md: `2px solid ${colorByKind[statusKind].border}`,
-                            },
-                            background: colorByKind[statusKind].bg,
-                            transition: "all .2s",
-                        }}
-                    >
-
-                    <Stack direction="row" spacing={1.5} alignItems="center" sx={{ mb: 1 }}>
-                            {statusKind === "ok" && <CheckCircleIcon color="success" />}
-                            {statusKind === "done" && <CheckCircleIcon color="info" />}
-                            {statusKind === "warn" && <WarningAmberIcon color="warning" />}
-                            {statusKind === "error" && <ErrorOutlineIcon color="error" />}
-                            <Typography variant="h6" fontWeight={800}>
-                                {statusKind === "ok"   && "ACCESSO CONSENTITO (auto +1)"}
-                                {statusKind === "done" && "GIÀ UTILIZZATO (0 rimanenti)"}
-                                {statusKind === "warn" && (result?.reason === "expired" ? "TOKEN SCADUTO" : "EVENTO NON DI OGGI")}
-                                {statusKind === "error"&& "NON VALIDO"}
-                                {statusKind === "idle" && "In attesa di scansione"}
-                            </Typography>
-                        </Stack>
-
-                        {result?.valid && (
-                            <>
-                                <Grid container spacing={1.5}>
-                                    <Grid item xs={12} md={6}>
-                                        <InfoRow label="Nome" value={`${result.nome || "—"} ${result.cognome || ""}`} />
-                                        <InfoRow label="Email" value={result.email || "—"} copy />
-                                        <InfoRow label="Telefono" value={result.telefono || "—"} copy />
-                                    </Grid>
-                                    <Grid item xs={12} md={6}>
-                                        <InfoRow label="Evento" value={ev?.name || result.eventId} />
-                                        <InfoRow label="Data evento" value={ev?.date || "—"} />
-                                        <InfoRow label="Biglietti" value={`${result.checkedInCount || 0} / ${result.quantity} (rim. ${result.remaining ?? 0})`} />
-                                    </Grid>
-                                </Grid>
-
-                                {/* Esito & azioni */}
-                                <Stack
-                                    direction={{ xs: "column", md: "row" }}
-                                    spacing={{ xs: 1, md: 1.5 }}
-                                    sx={{ mt: 2, width: "100%" }}
-                                >
-                                    <Button
-                                        fullWidth
-                                        variant="contained"
-                                        onClick={() => doCheckIn(result.token, 1)}
-                                        disabled={(result.remaining ?? 0) <= 0}
-                                        sx={{ height: 46, borderRadius: 999, fontWeight: 800 }}
-                                    >
-                                        +1 Check-in
-                                    </Button>
-
-                                    <Button
-                                        fullWidth
-                                        variant="outlined"
-                                        onClick={doUndo}
-                                        disabled={(result?.checkedInCount || 0) <= 0}
-                                        sx={{ height: 46, borderRadius: 999, fontWeight: 700 }}
-                                    >
-                                        Annulla ultimo
-                                    </Button>
-
-                                    <Button
-                                        fullWidth
-                                        variant="outlined"
-                                        startIcon={<LinkIcon />}
-                                        onClick={() => {
-                                            const base = (import.meta.env.VITE_API_BASE_URL || "http://localhost:3000").replace(/\/+$/, "");
-                                            window.open(`${base}/api/bookings/verify?token=${encodeURIComponent(result.token)}`,
-                                                "_blank", "noopener,noreferrer");
-                                        }}
-                                        sx={{ height: 46, borderRadius: 999, fontWeight: 700 }}
-                                    >
-                                        Apri verifica
-                                    </Button>
-                                </Stack>
-
-                            </>
-                        )}
-
-                        {!result && <Typography sx={{ opacity: .75 }}>Scansiona un QR o incolla un link/token per verificare.</Typography>}
-                    </Paper>
-                </Grid>
-            </Grid>
-
-            {/* MODAL OPERATORE FULL-SCREEN */}
-            <Dialog fullScreen open={opOpen} onClose={() => setOpOpen(false)}>
-                <Box sx={{ p: 2, background: "#0b0b0f", color: "#fff", minHeight: "100vh", display: "grid", gridTemplateRows: "auto 1fr auto", gap: 2 }}>
-                    <Stack direction="row" alignItems="center" justifyContent="space-between">
-                        <Button variant="text" startIcon={<FullscreenExitIcon />} onClick={() => setOpOpen(false)} sx={{ color: "#fff" }}>
-                            Chiudi
-                        </Button>
-                        <IconButton
-                            onClick={() => {
-                                const next = !soundOn;
-                                setSoundOn(next);
-                                try { localStorage.setItem("checkinSound", next ? "on" : "off"); } catch {}
-                            }}
-                            sx={{ color: "#fff" }}
-                        >
-                            {soundOn ? <VolumeUpIcon /> : <VolumeOffIcon />}
-                        </IconButton>
-
-                    </Stack>
-
-                    <Box sx={{ display: "grid", placeItems: "center" }}>
-                        <Box sx={{ textAlign: "center" }}>
-                            <Box sx={{
-                                width: 220, height: 220, borderRadius: "50%",
-                                mx: "auto", mb: 3,
-                                background:
-                                    statusKind === "ok"   ? "radial-gradient(circle, #1e7f39 0%, #0f3f20 70%)" :
-                                        statusKind === "done" ? "radial-gradient(circle, #1b4d74 0%, #0b2336 70%)" :
-                                            statusKind === "warn" ? "radial-gradient(circle, #8a6d1a 0%, #3d2f0a 70%)" :
-                                                statusKind === "error"? "radial-gradient(circle, #7f1e1e 0%, #3f0f0f 70%)" :
-                                                    "radial-gradient(circle, #333 0%, #111 70%)",
-                                boxShadow: "0 0 60px rgba(0,0,0,.6), inset 0 0 40px rgba(0,0,0,.5)",
-                                border: "6px solid rgba(255,255,255,.08)",
-                            }} />
-
-                            <Typography sx={{ fontSize: 40, fontWeight: 900, letterSpacing: 1, mb: 1 }}>
-                                {statusKind === "ok"   && "VALIDO"}
-                                {statusKind === "done" && "COMPLETO"}
-                                {statusKind === "warn" && (result?.reason === "expired" ? "SCADUTO" : "NON DI OGGI")}
-                                {statusKind === "error"&& "NON VALIDO"}
-                                {statusKind === "idle" && "IN ATTESA"}
-                            </Typography>
-
-                            {result?.valid && (
-                                <>
-                                    <Typography sx={{ fontSize: 24, opacity: .9, mb: .5 }}>{(result.nome || "") + " " + (result.cognome || "")}</Typography>
-                                    <Typography sx={{ fontSize: 20, opacity: .75, mb: 2 }}>{eventById(result.eventId)?.name || result.eventId}</Typography>
-                                    <Typography sx={{ fontSize: 22, fontWeight: 700 }}>
-                                        {`Entrati: ${result.checkedInCount || 0} / ${result.quantity} — Rim.: ${result.remaining ?? 0}`}
-                                    </Typography>
-                                </>
-                            )}
-                        </Box>
-                    </Box>
-
-                    {/* --- PATCH: bottoni footer modal operatore --- */}
-                    <Stack
-                        direction={{ xs: "column", sm: "row" }}
-                        spacing={1}
-                        justifyContent="center"
-                        sx={{ pb: 1, width: "100%" }}
-                        flexWrap={{ xs: "nowrap", sm: "wrap" }}
-                    >
-                        <Button
-                            variant="contained"
-                            size="large"
-                            onClick={startScan}
-                            startIcon={<QrCodeScannerIcon />}
-                            sx={{ width: { xs: "100%", sm: "auto" }, minHeight: 48 }}
-                        >
-                            Scansiona
-                        </Button>
-
-                        {result?.valid && (
-                            <>
-                                <Button
-                                    variant="contained"
-                                    size="large"
-                                    onClick={() => doCheckIn(result.token, 1)}
-                                    disabled={(result.remaining ?? 0) <= 0}
-                                    sx={{ width: { xs: "100%", sm: "auto" }, minHeight: 48 }}
-                                >
-                                    +1
-                                </Button>
-
-                                <Button
-                                    variant="outlined"
-                                    size="large"
-                                    onClick={doUndo}
-                                    disabled={(result?.checkedInCount || 0) <= 0}
-                                    sx={{ width: { xs: "100%", sm: "auto" }, minHeight: 48 }}
-                                >
-                                    Annulla
-                                </Button>
-                            </>
-                        )}
-                    </Stack>
-
-                </Box>
-            </Dialog>
-        </Box>
-    );
-}
-
-
-
-// Piccola riga info con bottone copia
-// ⬇️ SOSTITUISCI TUTTA InfoRow con questa
-function InfoRow({ label, value, copy = false }) {
-    return (
-        <Stack
-            direction="row"
-            spacing={1}
-            alignItems="center"
-            sx={{ mb: 0.6, minWidth: 0 }}
-        >
-            <Typography
-                sx={{
-                    flex: { xs: "0 0 96px", sm: "0 0 120px" },
-                    width: { xs: 96, sm: 120 },
-                    opacity: 0.75,
-                }}
-            >
-                {label}
-            </Typography>
-
-            <Typography
-                sx={{
-                    flex: "1 1 auto",
-                    minWidth: 0,
-                    fontWeight: 600,
-                    overflowWrap: "anywhere",   // <-- niente scroll orizzontale
-                    wordBreak: "break-word",
-                }}
-            >
-                {value}
-            </Typography>
-
-            {copy && (
-                <IconButton
-                    size="small"
-                    onClick={async () => {
-                        try { await navigator.clipboard.writeText(String(value || "")); } catch {}
-                    }}
-                    sx={{ flex: "0 0 auto", ml: 0.5, opacity: 0.8 }}
-                >
-                    <ContentCopyIcon fontSize="small" />
-                </IconButton>
-            )}
-        </Stack>
-    );
-}
-
-
 
 export default AdminPanel;
