+++ src/components/admin/MobileBookingCard.jsx
@@ -1,0 +1,19 @@
+import React from "react";
+import { Card, CardContent, Typography, Stack } from "@mui/material";
+
+function MobileBookingCard({ b }) {
+    return (
+        <Card sx={{ mb: 1.2, background: "#18181f", border: "1px solid rgba(255,255,255,0.06)" }}>
+            <CardContent>
+                <Typography variant="subtitle1" fontWeight={700}>{b.nome} {b.cognome}</Typography>
+                <Typography variant="body2" sx={{ opacity: 0.85 }}>{b.email}</Typography>
+                <Stack direction="row" spacing={2} sx={{ mt: 0.5, opacity: 0.85 }}>
+                    <Typography variant="caption">📞 {b.telefono}</Typography>
+                    <Typography variant="caption">🎟️ {b.quantity || 1}</Typography>
+                </Stack>
+            </CardContent>
+        </Card>
+    );
+}
+
+export default MobileBookingCard;
